# back/central-reserve/internal/infra/primary/http2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio REST para la gestión de reservas multi-restaurante.",
        "title": "Restaurant Reservation API",
        "termsOfService": "https://ejemplo.com/terminos",
        "contact": {
            "name": "Equipo de Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/business-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene la lista completa de tipos de negocio disponibles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BusinessType"
                ],
                "summary": "Obtener todos los tipos de negocio",
                "responses": {
                    "200": {
                        "description": "Tipos de negocio obtenidos exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeListSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo tipo de negocio con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BusinessType"
                ],
                "summary": "Crear tipo de negocio",
                "parameters": [
                    {
                        "description": "Datos del tipo de negocio",
                        "name": "businessType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BusinessTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tipo de negocio creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Código ya existe",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    }
                }
            }
        },
        "/business-types/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un tipo de negocio específico por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BusinessType"
                ],
                "summary": "Obtener tipo de negocio por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del tipo de negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de negocio obtenido exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tipo de negocio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un tipo de negocio existente con los datos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BusinessType"
                ],
                "summary": "Actualizar tipo de negocio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del tipo de negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del tipo de negocio",
                        "name": "businessType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BusinessTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de negocio actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tipo de negocio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Código ya existe",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina un tipo de negocio existente por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BusinessType"
                ],
                "summary": "Eliminar tipo de negocio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del tipo de negocio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de negocio eliminado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tipo de negocio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.BusinessTypeErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_infra_primary_http2_handlers_businesstypehandler_response.BusinessTypeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "request.BusinessTypeRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.BusinessTypeErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.BusinessTypeListResponse": {
            "type": "object",
            "properties": {
                "business_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_infra_primary_http2_handlers_businesstypehandler_response.BusinessTypeResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.BusinessTypeListSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.BusinessTypeListResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.BusinessTypeSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_infra_primary_http2_handlers_businesstypehandler_response.BusinessTypeResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ingrese su token JWT con el prefijo **Bearer**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}